+++ backend/bang-ggood/src/main/java/com/bang_ggood/category/dto/response/CategoryReadResponse.java
@@ -1,4 +1,4 @@
-package com.bang_ggood.category.dto;
+package com.bang_ggood.category.dto.response;
 
 import com.bang_ggood.category.domain.Category;
