--- backend/bang-ggood/src/main/java/com/bang_ggood/checklist/dto/ChecklistQuestionsResponse.java
@@ -1,7 +1,0 @@
-package com.bang_ggood.checklist.dto;
-
-import com.bang_ggood.category.dto.CategoryQuestionsResponse;
-import java.util.List;
-
-public record ChecklistQuestionsResponse(List<CategoryQuestionsResponse> categories) {
-}
