--- backend/bang-ggood/src/main/java/com/bang_ggood/checklist/dto/WrittenChecklistResponse.java
@@ -1,8 +1,0 @@
-package com.bang_ggood.checklist.dto;
-
-import com.bang_ggood.category.dto.WrittenCategoryQuestionsResponse;
-import com.bang_ggood.room.dto.WrittenRoomResponse;
-import java.util.List;
-
-public record WrittenChecklistResponse(WrittenRoomResponse room, List<Integer> options, List<WrittenCategoryQuestionsResponse> categories) {
-}
