+++ backend/bang-ggood/src/main/java/com/bang_ggood/checklist/dto/request/QuestionCreateRequest.java
@@ -1,4 +1,4 @@
-package com.bang_ggood.checklist.dto;
+package com.bang_ggood.checklist.dto.request;
 
 import jakarta.validation.constraints.NotNull;
