+++ backend/bang-ggood/src/main/java/com/bang_ggood/checklist/dto/response/CategoryScoreReadResponse.java
@@ -1,4 +1,4 @@
-package com.bang_ggood.checklist.dto;
+package com.bang_ggood.checklist.dto.response;
 
 public record CategoryScoreReadResponse(
         Integer categoryId, String categoryName, Integer score
