+++ backend/bang-ggood/src/main/java/com/bang_ggood/checklist/dto/response/ChecklistsComparisonReadResponse.java
@@ -1,4 +1,4 @@
-package com.bang_ggood.checklist.dto;
+package com.bang_ggood.checklist.dto.response;
 
 import java.util.List;
