+++ backend/bang-ggood/src/main/java/com/bang_ggood/checklist/dto/response/UserChecklistPreviewResponse.java
@@ -1,6 +1,7 @@
-package com.bang_ggood.checklist.dto;
+package com.bang_ggood.checklist.dto.response;
 
 import com.bang_ggood.checklist.domain.Checklist;
+
 import java.time.LocalDateTime;
 import java.util.List;
